Rename scripts/config.ts to scripts/hmy_config.ts for clarity
