+++ contract/relay/scripts/hmy_config.ts
